backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.122.0",
    "passlib[bcrypt]>=1.7.4",
    "argon2-cffi>=23.1.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlmodel>=0.0.27",
    "uvicorn>=0.38.0",
    "python-dotenv>=1.0.0",
    "alembic>=1.11.0",
    "psycopg[binary]>=3.1",
]
